+++ processor/src/test/resources/expected/com/example/repository/DaggerEnabledRepositoryDaggerModule.java
@@ -1,0 +1,18 @@
+package com.example.repository;
+
+import dagger.Module;
+import dagger.Provides;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.inject.Singleton;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+@Module
+public interface DaggerEnabledRepositoryDaggerModule {
+  @Nonnull
+  @Singleton
+  @Provides
+  static DaggerEnabledRepositoryRepository repository() {
+    return DaggerEnabledRepositoryRepository.newRepository();
+  }
+}