+++ processor/src/test/resources/input/com/example/repository/DaggerEnabledRepository.java
@@ -1,0 +1,31 @@
+package com.example.repository;
+
+import javax.annotation.Nonnull;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Observable;
+import org.realityforge.arez.annotations.Repository;
+
+@Repository( dagger = true )
+@ArezComponent
+public class DaggerEnabledRepository
+{
+  @Nonnull
+  private String _name;
+
+  DaggerEnabledRepository( @Nonnull final String name )
+  {
+    _name = name;
+  }
+
+  @Observable
+  @Nonnull
+  public String getName()
+  {
+    return _name;
+  }
+
+  public void setName( @Nonnull final String name )
+  {
+    _name = name;
+  }
+}